Unit tests & circular dependency fix
