+++ i18n/sample/src/profiles/utils.i18n.json
@@ -1,0 +1,22 @@
+{
+    "collectProfileDetails.profileTypeMissing": "No profile type was chosen. Operation Cancelled",
+    "collectProfileDetails.option.prompt.url.placeholder": "url:port",
+    "collectProfileDetails.option.prompt.url": "Enter a z/OS URL in the format 'url:port'.",
+    "collectProfileDetails.invalidzosURL": "Please enter a valid host URL in the format 'url:port'.",
+    "collectProfileDetails.zosmfURL": "No valid value for z/OS URL. Operation Cancelled",
+    "collectProfileDetails.invalidPort": "Please enter a valid port number",
+    "collectProfileDetails.option.prompt.port.placeholder": "Port Number",
+    "collectProfileDetails.undefined.port": "Invalid Port number provided or operation was cancelled",
+    "collectProfileDetails.option.prompt.username.placeholder": "Optional: User Name",
+    "collectProfileDetails.option.prompt.username": "Enter the user name for the connection.",
+    "collectProfileDetails.invalidUser": "Please enter a valid password",
+    "collectProfileDetails.undefined.user": "Invalid user provided or operation was cancelled",
+    "collectProfileDetails.undefined.username": "No username defined.",
+    "collectProfileDetails.option.prompt.password.placeholder": "Optional: Password",
+    "collectProfileDetails.option.prompt.password": "Enter the password for the connection.",
+    "collectProfileDetails.undefined.pass": "Invalid password provided or operation was cancelled",
+    "collectProfileDetails.undefined.password": "No password defined.",
+    "collectProfileDetails.option.prompt.ru.placeholder": "Reject Unauthorized Connections",
+    "collectProfileDetails.rejectUnauthorize": "No certificate option selected. Operation Cancelled",
+    "collectProfileDetails.booleanValue": "No boolean selected. Operation Cancelled"
+}
