+++ src/api/ZoweExplorerZosmfApi.ts
@@ -13,9 +13,10 @@ import * as zowe from "@zowe/cli";
 import { ICommandArguments, ConnectionPropsForSessCfg, ISession, Session, IProfileLoaded, ITaskWithStatus } from "@zowe/imperative";
 import { ZoweExplorerApi } from "./ZoweExplorerApi";
 import * as nls from "vscode-nls";
-import { errorHandling } from "../utils";
 import { DefaultProfileManager } from "../profiles/DefaultProfileManager";
-import { Profiles } from "../Profiles";
+import { collectProfileDetails } from "../profiles/utils";
+import * as vscode from "vscode";
+import * as globals from "../globals";
 
 // Set up localization
 nls.config({ messageFormat: nls.MessageFormat.bundle, bundleFormat: nls.BundleFormat.standalone })();
@@ -72,6 +73,8 @@ class ZosmfApiCommon implements ZoweExplorerApi.ICommon {
         const baseProfile = DefaultProfileManager.getInstance().getDefaultProfile("base");
 
         // If user exists in serviceProfile, use serviceProfile to login because it has precedence over baseProfile
+        // If no user & no baseProfile, use serviceProfile as default
+        // If no user & no token, use serviceProfile as default
         if (serviceProfile.profile.user || !baseProfile || (!serviceProfile.profile.user && baseProfile && !baseProfile.profile.tokenValue)) {
             if (prompt) {
                 // Select for prompting only fields which are not defined
@@ -91,9 +94,25 @@ class ZosmfApiCommon implements ZoweExplorerApi.ICommon {
                 }
 
                 try {
-                    const newDetails = await Profiles.getInstance().collectProfileDetails(schemaArray);
+                    const newDetails = await collectProfileDetails(schemaArray);
                     for (const detail of schemaArray) { serviceProfile.profile[detail] = newDetails[detail]; }
-                } catch (error) { await errorHandling(error); }
+                } catch (error) {
+                    // tslint:disable:no-magic-numbers
+                    if (error.mDetails && error.mDetails.errorCode === 401) {
+                        if (globals.ISTHEIA) {
+                            vscode.window.showErrorMessage(error.message);
+                            this.getValidSession(serviceProfile, serviceProfile.name, true);
+                        } else {
+                            vscode.window.showErrorMessage(error.message, "Check Credentials").then(async (selection) => {
+                                if (selection) {
+                                    delete serviceProfile.profile.user;
+                                    delete serviceProfile.profile.password;
+                                    await this.getValidSession(serviceProfile, serviceProfile.name, true);
+                                }
+                            });
+                        }
+                    } else { throw error; }
+                }
             }
             const cmdArgs: ICommandArguments = {
                 $0: "zowe",
@@ -119,7 +138,23 @@ class ZosmfApiCommon implements ZoweExplorerApi.ICommon {
             } catch (error) {
                 // When no password is entered, we should silence the error message for not providing it
                 // since password is optional in Zowe Explorer
-                if (error.message !== "Must have user & password OR base64 encoded credentials") { await errorHandling(error); }
+                if (error.message !== "Must have user & password OR base64 encoded credentials") {
+                    // tslint:disable:no-magic-numbers
+                    if (error.mDetails && error.mDetails.errorCode === 401) {
+                        if (globals.ISTHEIA) {
+                            vscode.window.showErrorMessage(error.message);
+                            this.getValidSession(serviceProfile, serviceProfile.name, true);
+                        } else {
+                            vscode.window.showErrorMessage(error.message, "Check Credentials").then(async (selection) => {
+                                if (selection) {
+                                    delete serviceProfile.profile.user;
+                                    delete serviceProfile.profile.password;
+                                    await this.getValidSession(serviceProfile, serviceProfile.name, true);
+                                }
+                            });
+                        }
+                    } else { throw error; }
+                }
             }
         } else if (baseProfile) {
             // baseProfile exists, so APIML login is possible
@@ -145,7 +180,7 @@ class ZosmfApiCommon implements ZoweExplorerApi.ICommon {
                         cmdArgs,
                         { requestToken: false,
                           doPrompting: prompt,
-                          getValuesBack: Profiles.getInstance().collectProfileDetails });
+                          getValuesBack: collectProfileDetails });
                 } else {
                     connectableSessCfg = await ConnectionPropsForSessCfg.addPropsOrPrompt<ISession>(sessCfg,
                                                                                                     cmdArgs,
@@ -154,7 +189,22 @@ class ZosmfApiCommon implements ZoweExplorerApi.ICommon {
 
                 return new Session(connectableSessCfg);
             } catch (error) {
-                await errorHandling(error); }
+                // tslint:disable:no-magic-numbers
+                if (error.mDetails && error.mDetails.errorCode === 401) {
+                    if (globals.ISTHEIA) {
+                        vscode.window.showErrorMessage(error.message);
+                        this.getValidSession(serviceProfile, serviceProfile.name, true);
+                    } else {
+                        vscode.window.showErrorMessage(error.message, "Check Credentials").then(async (selection) => {
+                            if (selection) {
+                                delete serviceProfile.profile.user;
+                                delete serviceProfile.profile.password;
+                                await this.getValidSession(serviceProfile, serviceProfile.name, true);
+                            }
+                        });
+                    }
+                } else { throw error; }
+            }
         } else {
             // Neither baseProfile nor serviceProfile exists. It is impossible to login with the currently-provided information.
             throw new Error(localize("getValidSession.loginImpossible",
