+++ src/job/ZoweJobNode.ts
@@ -16,12 +16,12 @@ import { Session, IProfileLoaded } from "@zowe/imperative";
 import { IZoweJobTreeNode } from "../api/IZoweTreeNode";
 import { ZoweTreeNode } from "../abstract/ZoweTreeNode";
 import { errorHandling } from "../utils";
-import { refreshTree } from "../profiles/utils";
 import { ZoweExplorerApiRegister } from "../api/ZoweExplorerApiRegister";
 import { getIconByNode } from "../generators/icons";
 import * as contextually from "../shared/context";
 
 import * as nls from "vscode-nls";
+import { refreshTree } from "../shared/utils";
 // Set up localization
 nls.config({ messageFormat: nls.MessageFormat.bundle, bundleFormat: nls.BundleFormat.standalone })();
 const localize: nls.LocalizeFunc = nls.loadMessageBundle();
